AiReality2Vars.Main (Plugin: ai.reality2.vars v0.1.0)

Module for managing the main supervisor tree for the AiReality2Vars App.

In this instance, the main supervisor is a DynamicSupervisor, which is used to manage the Data stores for each Sentant, however, this could be a PartitionSupervisor, or indeed, just a single process for all Sentants.

Use this as a template for your own Main module for your own Apps. The create, delete, sendto and whereas functions must be implemented.

Author

Dr. Roy C. Davies

roycdavies.github.io

Functions

create(sentant_id)

@spec create(String.t()) :: {:ok} | {:error, :existance}

Create a new Data store, returning {:ok} or an appropriate error.

This creates a new child for each Sentant where the id is passed in and has the App name appended.

Parameters

id - The id of the Sentant for which the Data store is being created.

delete(sentant_id)

@spec delete(String.t()) :: {:ok} | {:error, :existance}

Delete a Data store, returning {:ok} or an appropriate error.

Parameters

id - The id of the Sentant for which the Data store is being deleted.

sendto(sentant_id, command_and_parameters)

@spec sendto(String.t(), map()) :: {:ok} | {:error, :unknown_command}

Send a command to the Data store for the given Sentant id.

Depending on the command, this might be a synchronous call or an asynchronous cast. Ideally, use this convention when creating your own Apps so that the commands are consistent across all Apps.

Parameters

id - The id of the Sentant for which the command is being sent.

command - A map containing the command and parameters to be sent.

Returns

{:ok} - If the command was sent successfully.

{:error, :unknown_command} - If the command was not recognised.

whereis(sentant_id)

@spec whereis(String.t()) :: pid() | nil

Return the process id that can be used for subsequent communications.

In this implementation, each Sentant gets its own Data store GenServer process, but other Apps might just use a single process for all Sentants. Externally this is transparent.

Parameters

id - The id of the Sentant for which process id is being returned.